Move default reader to separate file
